[fix] Fixed an issue where some emojis were not converted
I found that there are some emojis, such as '🥳 ', that are not judged
as emojis in the unic-emoji-char crate. Therefore, I changed the crate
used to any-ascii crate. This crate can recognize such emojis, and since
a colon is added at the beginning and the end of the conversion, it is
now possible to determine whether or not a emoji is a emoji. This crate
also eliminates the need for deunicode.