[change] Changed to construct `Input` from data in memory
The "Too many open files" error was occurring, and it was thought that
this was caused by stdin not being closed, so I changed to build the
`Input` structure after the writing from stdin to the ffmpeg process was
completed.
In addition, the same error can still occur in the future, so I added a
log output to supplement the error.