[change] Changed to generate `Input` without ffmpeg
I had abandoned this approach due to the problem of noise at the
beginning of the reading whenever I tried to generate `Input` without
ffmpeg, but it turned out that the noise was caused by reading the
metadata as audio data as well. Also, while the audio data obtained by
synthesis was in pcm_s16le format, the `Reader` of `Input' required
`Vec<u8>`, so it had to be converted in little-endian format.

close #2